#60 UserModelを追加しGymとTrainerはUserModelを継承するようにする
