+++ src/app/Models/Gym.php
@@ -2,9 +2,7 @@
 
 namespace App\Models;
 
-use Illuminate\Database\Eloquent\Model;
-
-class Gym extends Model
+class Gym extends User
 {
     protected $fillable = [
         'president_name',
@@ -17,25 +15,4 @@ class Gym extends Model
     protected $casts = [
         'requirements' => 'json',
     ];
-
-    public function login()
-    {
-        return $this->morphOne(Login::class, 'user');
-    }
-
-    public function matchingCondition()
-    {
-        return $this->morphOne(MatchingCondition::class, 'user');
-    }
-
-    /**
-     * ログインと関連付ける
-     * @param int $id
-     * @return \App\Models\Login
-     */
-    public function associateToLogin(Login $login, array $updateColumns = [])
-    {
-        // 更新するカラムとマージ
-        return $this->login()->save($login->fillUpdateColumns($updateColumns));
-    }
 }
