+++ src/app/Models/Trainer.php
@@ -2,9 +2,7 @@
 
 namespace App\Models;
 
-use Illuminate\Database\Eloquent\Model;
-
-class Trainer extends Model
+class Trainer extends User
 {
     protected $fillable = [
         'tel',
@@ -12,25 +10,4 @@ class Trainer extends Model
     ];
 
     protected $with = ['login'];
-
-    public function login()
-    {
-        return $this->morphOne(Login::class, 'user');
-    }
-
-    public function matchingCondition()
-    {
-        return $this->morphOne(MatchingCondition::class, 'user');
-    }
-
-    /**
-     * ログインと関連付ける
-     * @param int $id
-     * @return \App\Models\Login
-     */
-    public function associateToLogin(Login $login, array $updateColumns = [])
-    {
-        // 更新するカラムとマージ
-        return $this->login()->save($login->fillUpdateColumns($updateColumns));
-    }
 }
